Rework ResolvedSourcePackage (take one)
This is good but ...

now I reached

readRepoIndex
  :: Verbosity
  -> RepoContext
  -> Repo
  -> RepoIndexState
  -> IO (PackageIndex UnresolvedSourcePackage, [Dependency], IndexStateInfo)

which is where the change has to be. UnresolvedSourcePackage there is
correct, source packages from a repository are not downloaded right away
so they don't have a local path.

BUT this is where we should be injecting the tarball hash since it comes
from the repo.

This file is a mess, it should be heavily refactored!